Fixes on improvement
