+++ requirements.txt
@@ -61,3 +61,10 @@ validators==0.22.0
 watchdog==4.0.0
 zipp==3.17.0
 sublist3r
+asyncio
+httpx
+trafilatura
+os
+urllib.parse
+time
+testgoogleapi
